Move JsonReader to Study updater where it's used
