+++ src/cpp/helpers/CMakeLists.txt
@@ -23,7 +23,6 @@ target_sources(helpers PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/ArchiveWriter.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/Clock.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/FileInBuffer.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/JsonXpansionReader.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/OptionsParser.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/helpers/AntaresArchiveUpdater.h
         ${CMAKE_CURRENT_SOURCE_DIR}/include/antares-xpansion/helpers/AntaresArchiveUpdaterExeOptions.h
