+++ tests/cpp/CMakeLists.txt
@@ -32,3 +32,4 @@ add_subdirectory(zip_mps)
 add_subdirectory(benders)
 add_subdirectory(full_run)
 add_subdirectory(outer_loop)
+add_subdirectory(study_updater)
